+++ modules/alloy/src/main/ts/ephox/alloy/dragging/dragndrop/ImageClone.ts
@@ -1,4 +1,4 @@
-import { Arr, Cell, Optional } from '@ephox/katamari';
+import { Arr, Singleton } from '@ephox/katamari';
 import { PlatformDetection } from '@ephox/sand';
 import { Attribute, Css, DomEvent, Insert, Remove, Replication, SugarElement } from '@ephox/sugar';
 
@@ -66,7 +66,7 @@ const blockDefaultGhost = (target: SugarElement): void => {
 // Edge doesn't have setDragImage support and just hiding the target element will position the drop icon incorrectly so we need custom ghost
 // TODO: Get rid of this once Edge switches to Chromium we feature detect setDragImage support so once they have it should use that instead
 const setDragImageFromCloneEdgeFallback = (image: DragnDropImageClone, parent: SugarElement, target: SugarElement): void => {
-  const ghostState = Cell(Optional.none<SugarElement<any>>());
+  const ghostState = Singleton.value<SugarElement<any>>();
 
   const drag = DomEvent.bind(target, 'drag', (evt) => {
     // The calculated position needs to at least be cord + 1 since it would otherwise interfere with dropping
@@ -76,7 +76,7 @@ const setDragImageFromCloneEdgeFallback = (image: DragnDropImageClone, parent: S
     const ghost = ghostState.get().getOrThunk(() => {
       const newGhost = createGhostClone(image);
 
-      ghostState.set(Optional.some(newGhost));
+      ghostState.set(newGhost);
 
       Css.setAll(newGhost, {
         position: 'fixed',
@@ -98,7 +98,7 @@ const setDragImageFromCloneEdgeFallback = (image: DragnDropImageClone, parent: S
   });
 
   const dragEnd = DomEvent.bind(target, 'dragend', (_) => {
-    ghostState.get().each(Remove.remove);
+    ghostState.on(Remove.remove);
     drag.unbind();
     dragEnd.unbind();
   });