+++ modules/tinymce/src/core/main/ts/undo/Operations.ts
@@ -5,8 +5,6 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Optional } from '@ephox/katamari';
-
 import Editor from '../api/Editor';
 import * as Settings from '../api/Settings';
 import Tools from '../api/util/Tools';
@@ -18,7 +16,7 @@ import { Index, Locks, UndoBookmark, UndoLevel, UndoManager } from './UndoManage
 
 export const beforeChange = (editor: Editor, locks: Locks, beforeBookmark: UndoBookmark) => {
   if (isUnlocked(locks)) {
-    beforeBookmark.set(Optional.some(GetBookmark.getUndoBookmark(editor.selection)));
+    beforeBookmark.set(GetBookmark.getUndoBookmark(editor.selection));
   }
 };